Extract out some helpers to separate files
