+++ contracts/service-provider-directory/src/test_helpers/mod.rs
@@ -1,4 +1,6 @@
 pub mod assert;
 pub mod fixture;
 pub mod helpers;
+pub mod signing;
+pub mod transactions;
 //pub mod test_setup;
